Adopt commit 072727b19 from master: Merge pull request #1029 from hrosen4/bait-and-switch
commit 072727b192fab48793e18eaa634a51eb81a69c09
Merge: 6ca945629 db1881984
Author: Henry Verschell <hverschell@bloomberg.net>
Date:   Thu Mar 16 22:16:42 2017 -0400

    Merge pull request #1029 from hrosen4/bait-and-switch

    The type for Switch SELECTOR should consistently be unsigned.

commit db188198418f29c8011d6c008ed8b02531a919d7
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Tue Mar 14 19:00:20 2017 -0400

    Use size_t for SELECTOR type.

commit 4e2e482a06ab64c40e3aea69ea32683d696a97b6
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Thu Mar 9 11:51:18 2017 -0500

    The type for Switch SELECTOR should consistently be unsigned.

+++ groups/bsl/bslmf/bslmf_switch.t.cpp
@@ -340,7 +340,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch2<2,A,B>::Type()));
             ASSERT('0' == f(bslmf::Switch2<10,A,B>::Type()));
             ASSERT('0' ==
-                 f(bslmf::Switch2<static_cast<unsigned int>(-1),A,B>::Type()));
+                   f(bslmf::Switch2<static_cast<size_t>(-1), A, B>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch3.\n");
@@ -354,7 +354,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch3<3,A,B,C>::Type()));
             ASSERT('0' == f(bslmf::Switch3<10,A,B,C>::Type()));
             ASSERT('0' ==
-               f(bslmf::Switch3<static_cast<unsigned int>(-1),A,B,C>::Type()));
+                   f(bslmf::Switch3<static_cast<size_t>(-1),A,B,C>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch4.\n");
@@ -369,7 +369,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch4<4,A,B,C,D>::Type()));
             ASSERT('0' == f(bslmf::Switch4<10,A,B,C,D>::Type()));
             ASSERT('0' ==
-             f(bslmf::Switch4<static_cast<unsigned int>(-1),A,B,C,D>::Type()));
+                   f(bslmf::Switch4<static_cast<size_t>(-1),A,B,C,D>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch5.\n");
@@ -385,7 +385,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch5<5,A,B,C,D,E>::Type()));
             ASSERT('0' == f(bslmf::Switch5<10,A,B,C,D,E>::Type()));
             ASSERT('0' ==
-           f(bslmf::Switch5<static_cast<unsigned int>(-1),A,B,C,D,E>::Type()));
+                 f(bslmf::Switch5<static_cast<size_t>(-1),A,B,C,D,E>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch6.\n");
@@ -402,7 +402,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch6<6,A,B,C,D,E,F>::Type()));
             ASSERT('0' == f(bslmf::Switch6<10,A,B,C,D,E,F>::Type()));
             ASSERT('0' ==
-         f(bslmf::Switch6<static_cast<unsigned int>(-1),A,B,C,D,E,F>::Type()));
+               f(bslmf::Switch6<static_cast<size_t>(-1),A,B,C,D,E,F>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch7.\n");
@@ -420,7 +420,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch7<7,A,B,C,D,E,F,G>::Type()));
             ASSERT('0' == f(bslmf::Switch7<10,A,B,C,D,E,F,G>::Type()));
             ASSERT('0' ==
-       f(bslmf::Switch7<static_cast<unsigned int>(-1),A,B,C,D,E,F,G>::Type()));
+             f(bslmf::Switch7<static_cast<size_t>(-1),A,B,C,D,E,F,G>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch8.\n");
@@ -439,7 +439,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch8<8,A,B,C,D,E,F,G,H>::Type()));
             ASSERT('0' == f(bslmf::Switch8<10,A,B,C,D,E,F,G,H>::Type()));
             ASSERT('0' ==
-     f(bslmf::Switch8<static_cast<unsigned int>(-1),A,B,C,D,E,F,G,H>::Type()));
+           f(bslmf::Switch8<static_cast<size_t>(-1),A,B,C,D,E,F,G,H>::Type()));
         }
 
         if (verbose) printf("\tTesting 'bslmf_Switch9.\n");
@@ -459,7 +459,7 @@ int main(int argc, char *argv[])
             ASSERT('0' == f(bslmf::Switch9<9,A,B,C,D,E,F,G,H,I>::Type()));
             ASSERT('0' == f(bslmf::Switch9<10,A,B,C,D,E,F,G,H,I>::Type()));
             ASSERT('0' ==
-   f(bslmf::Switch9<static_cast<unsigned int>(-1),A,B,C,D,E,F,G,H,I>::Type()));
+         f(bslmf::Switch9<static_cast<size_t>(-1),A,B,C,D,E,F,G,H,I>::Type()));
         }
 #endif  // BSLMF_SWITCH_TEST_SWITCH_WITH_NAMED_LENGTH
       } break;
@@ -505,7 +505,7 @@ int main(int argc, char *argv[])
         ASSERT('0' == f(bslmf::Switch<10,A,B,C,D,E,F,G,H,I,J>::Type()));
         ASSERT('0' == f(bslmf::Switch<100,A,B,C,D,E,F,G,H,I,J>::Type()));
         ASSERT('0' ==
-  f(bslmf::Switch<static_cast<unsigned int>(-1),A,B,C,D,E,F,G,H,I,J>::Type()));
+        f(bslmf::Switch<static_cast<size_t>(-1),A,B,C,D,E,F,G,H,I,J>::Type()));
 
         if (verbose) printf("Testing Default arguments\n");
         ASSERT('A' == f(bslmf::Switch<0,A>::Type()));
@@ -518,8 +518,7 @@ int main(int argc, char *argv[])
         ASSERT('0' == f(bslmf::Switch<7,A>::Type()));
         ASSERT('0' == f(bslmf::Switch<8,A>::Type()));
         ASSERT('0' == f(bslmf::Switch<10,A>::Type()));
-        ASSERT('0' ==
-               f(bslmf::Switch<static_cast<unsigned int>(-1),A>::Type()));
+        ASSERT('0' == f(bslmf::Switch<static_cast<size_t>(-1),A>::Type()));
         ASSERT('B' == f(bslmf::Switch<1,A,B>::Type()));
         ASSERT('0' == f(bslmf::Switch<2,A,B>::Type()));
